+++ CHANGELOG.rst
@@ -68,6 +68,10 @@ New Features
       >>> skill.sst.plot(hue="model", col="month", col_wrap=3)
 
   (:issue:`635`, :pr:`690`) `Aaron Spring`_.
+- :py:meth:`~climpred.classes.HindcastEnsemble.plot_alignment` shows how forecast and observations are
+  aligned based on the `alignment <alignment.html>`_ keyword. This may help
+  understanding which dates are matched for the different ``alignment`` approaches.
+  (:issue:`701`, :pr:`702`) `Aaron Spring`_.
 - Add ``attrs`` to new ``coordinates`` created by ``climpred``. (:issue:`695`, :pr:`697`) `Aaron Spring`_.
 
 Internals/Minor Fixes